requirements.txt
aiomysql==0.2.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bokeh==2.4.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
dateparser==1.2.0
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
feedparser==6.0.11
fonttools==4.53.1
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
llvmlite==0.39.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.56.4
numpy==1.23.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parse==1.20.2
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-html==0.10.0
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.1
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlmodel==0.0.21
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
vectorbt==0.26.2
w3lib==2.2.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
widgetsnbextension==4.0.11
xyzservices==2024.6.0
yfinance==0.2.40
zipp==3.19.2
